[fileconsumer] Deduplicate fingerprints less aggressively (#24235)
When finding files, we would check for files with duplicate fingerprints
and deduplicate them if `a.StartsWith(b)` or `b.StartsWith(a)`. This
`StartsWith` logic is useful for recognizing files that have new content
since the previous poll interval. However, when deduplicating files
observed at the same moment, the only case that we need to consider is
copy/truncate rotation. In this case, a file may have been copied but
the original has not yet been truncated. At that moment we should expect
to find two files with exactly the same fingerprint. Therefore, we do
not need to check StartsWith cases.